chore: move SubGroupInput from components/refactor to components
