+++ src/modules/JobSkills.jsx
@@ -5,7 +5,7 @@ import Form from '../common/components/Form';
 import { v4 as uuidv4 } from 'uuid';
 import { turnArray } from "../common/components/helper/turnArray";
 import { turnGroupObject } from "../common/components/helper/turnGroupObject";
-import SubGroupInput from "../common/components/refactor/SubGroupInput";
+import SubGroupInput from "../common/components/SubGroupInput";
 import JobSkillsCard from "./JobSkillsCard";
 
 const JobSkills = () => {
